Moved `Source` out of photon.py
